fix: rename verseUtils
