+++ src/components/verseUtils.js
